ElasticNet-NSGAII-FS/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
PyYAML>=6.0
matplotlib>=3.7
scipy>=1.10
